requirements.txt
aiofiles==0.8.0
fastapi==0.79.0
monai==0.8.1
nibabel==3.2.2
numpy
#openvino==2022.1.0
openvino_dev
Pillow==9.2.0
pydantic==1.9.1
pytorch_lightning==1.6.3
PyYAML==6.0
scipy==1.8.1
setuptools==62.1.0
torch==1.11.0
